Created mouse handlers for press and release
It was decided to not use the mouse clicked event since clicks will be reset after each frame, if a click was registered in the middle of a frame it would not be reflected in the program. Instead, the presses and releases between click resets will be stored and then clicks will be calculated after the resets.