Updated AI with final parameters
After running a second, more extensive search, with a larger initial search and more trials per set of parameters (so there is less randomness involved and less likely to ignore a range of numbers after the first trial), I compared the results of this new search with the first search, this time with 50 trials per set and a larger limit of 10 000 minos. This showed a lot of variance which makes me think that a lot of the progress during the searches was simply due to randomness. However, the second search did provide better results than the first despite this. I considerred using a standard set of tetrominos to compare them, but then this performance would not translate to the actual application, where a similar variance would occur once it is attempted with random tetrominos.